added middleware, permissions
