+++ .gitignore
@@ -19,3 +19,6 @@
 
 # Go workspace file
 go.work
+
+# Env files
+*.env
