--- db/tests/main_test.go
@@ -1,29 +1,0 @@
-package db_test
-
-import (
-	"database/sql"
-	"log"
-	"os"
-	"testing"
-
-	db "github.com/olartbaraq/spectrumshelf/db/sqlc"
-	"github.com/olartbaraq/spectrumshelf/utils"
-)
-
-var testQueries *db.Queries
-
-func TestMain(m *testing.M) {
-	// This function is to perform the main test.
-
-	config, err := utils.LoadConfig("../..")
-	if err != nil {
-		log.Fatal("Could not load env config", err)
-	}
-
-	conn, err := sql.Open(config.DBdriver, config.DBsource)
-	if err != nil {
-		log.Fatalf("There was an error connecting to database: %v", err)
-	}
-	testQueries = db.New(conn)
-	os.Exit(m.Run())
-}
